+++ peek_plugin_graphdb/_private/storage/GraphDbSegment.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Index, ForeignKey
+from sqlalchemy import Column, Index, ForeignKey, BigInteger
 from sqlalchemy import Integer, String
 from sqlalchemy.orm import relationship
 from vortex.Tuple import Tuple, addTupleType
@@ -14,7 +14,7 @@ class GraphDbSegment(Tuple, DeclarativeBase):
     __tablename__ = 'GraphDbSegment'
 
     #:  The unique ID of this segment (database generated)
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
 
     #:  The model set for this segment
     modelSetId = Column(Integer,
